Rename result -> winner
