+++ src/main/java/com/arcbees/bourseje/client/admin/winner/WinnerView.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.arcbees.bourseje.client.admin.result;
+package com.arcbees.bourseje.client.admin.winner;
 
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.user.client.ui.Widget;
@@ -22,12 +22,12 @@ import com.gwtplatform.mvp.client.ViewImpl;
 
 import javax.inject.Inject;
 
-public class ResultView extends ViewImpl implements ResultPresenter.MyView {
-    interface Binder extends UiBinder<Widget, ResultView> {
+public class WinnerView extends ViewImpl implements WinnerPresenter.MyView {
+    interface Binder extends UiBinder<Widget, WinnerView> {
     }
 
     @Inject
-    ResultView(
+    WinnerView(
             Binder uiBinder) {
         initWidget(uiBinder.createAndBindUi(this));
     }
